pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-conversion-engine"
version = "0.1.0"
description = "RAG knowledge base document normalization and conversion service engine"
authors = [{ name = "RAG Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "celery>=5.3",
    "flower>=2.0",
    "redis>=5.0",
    "minio>=7.1",
    "prometheus-client>=0.17",
    "python-multipart>=0.0.7",
    "aiofiles>=23.2",
    "structlog>=23.1",
    "orjson>=3.9",
    "pypdf>=4.3",
    "requests>=2.31",
    "markdownify>=0.12",
    "gradio>=4.36",
    "openpyxl>=3.1",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-html>=4.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "ruff>=0.6"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
